requirements.txt
sqlalchemy>=2.0.0
streamlit>=1.30.0
anthropic>=0.18.0
pyyaml>=6.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
